+++ package.json
@@ -32,7 +32,7 @@
       "name": "Patrick Dwyer"
     }
   ],
-  "main": "index.js",
+  "main": "src/index.js",
   "bin": {
     "cyclonedx-bom": "./bin/make-bom.js",
     "cyclonedx-node": "./bin/make-bom.js"
@@ -66,9 +66,8 @@
     "xmlbuilder": "^15.1.1"
   },
   "files": [
-    "index.js",
+    "src/",
     "bin/",
-    "model/",
     "res/"
   ],
   "devDependencies": {
