+++ src/model/AttachmentText.js
