+++ src/model/Component.js
