+++ src/model/CycloneDXObject.js
