+++ src/model/Dependency.js
