+++ src/model/ExternalReference.js
