+++ src/model/ExternalReferenceList.js
