+++ src/model/Hash.js
