+++ src/model/HashList.js
