+++ src/model/License.js
