+++ src/model/Swid.js
