+++ tests/model/OrganizationalContact.test.js
@@ -19,7 +19,7 @@
  * Copyright (c) OWASP Foundation. All Rights Reserved.
  */
 
-const OrganizationalContact = require('../../model/OrganizationalContact')
+const OrganizationalContact = require('../../src/model/OrganizationalContact')
 
 test('Model: OrganizationalContact / Format: XML', () => {
   const contact = new OrganizationalContact('John Doe', 'john.doe@examp.com', '555-1212')
